x3dom - Viewer and Toolbox
This is a test viewer implemented in the www.rockartdatabase.com
project to display and help with basic analysis of 2D images and 3D
models. The tool allows the user to place markers, measure distances,
identify colour schemes, identify colours using a native colour picker
and adjust colour in images and 3D models for color frequency spectrum
analysis.